Fix flow-field infinite loop, straggler zombies, and UX polish
- Dijkstra distances now stored in Float64Array: float32 round-up made the
  same relaxation an 'improvement' forever, hanging the sim with large wall
  counts
- Horde spawns only placed where the colony is reachable; aggro zombies off
  the flow field shamble straight at the HQ instead of going passive
- Crowd separation no longer shoves zombies onto unwalkable tiles (dense
  hordes could ferry each other across water onto unreachable shores);
  marooned wave zombies relocate to a valid spawn edge so the final wave
  can always be cleared and victory verified end-to-end
- Stuck aggro zombies take a short wander burst to escape terrain notches
- HQ produces +2 food so the first tents don't starve the colony
- Selection panel no longer rebuilds every frame (demolish button was
  unclickable mid-refresh); crosshair cursor in build mode; vignette

+++ src/flowfield.js
@@ -9,8 +9,11 @@ export class FlowField {
   constructor(map) {
     this.map = map;
     const n = map.size * map.size;
-    this.dist = new Float32Array(n);
-    this.cost = new Float32Array(n);
+    // dist MUST be float64: storing float64 relaxations into a Float32Array
+    // can round *up*, making the same relaxation an "improvement" forever —
+    // an infinite Dijkstra loop.
+    this.dist = new Float64Array(n);
+    this.cost = new Float64Array(n);
   }
 
   // occ: Int32Array of building ids (+1) per tile, 0 = empty.
